+++ lib/array/strings.go
@@ -7,9 +7,6 @@ import (
 	"strings"
 
 	"github.com/artie-labs/transfer/lib/stringutil"
-
-	"github.com/artie-labs/transfer/lib/config/constants"
-	"github.com/artie-labs/transfer/lib/typing"
 )
 
 func InterfaceToArrayString(val interface{}) ([]string, error) {
@@ -67,52 +64,6 @@ func StringsJoinAddPrefix(args StringsJoinAddPrefixArgs) string {
 	}
 
 	return strings.Join(retVals, args.Separator)
-}
-
-// ColumnsUpdateQuery takes:
-// columns - list of columns to iterate
-// columnsToTypes - given that list, provide the types (separate list because this list may contain invalid columns
-// bigQueryTypeCasting - We'll need to escape the column comparison if the column's a struct.
-// It then returns a list of strings like: cc.first_name=c.first_name,cc.last_name=c.last_name,cc.email=c.email
-func ColumnsUpdateQuery(columns []string, columnsToTypes typing.Columns, bigQueryTypeCasting bool) string {
-	var _columns []string
-	for _, column := range columns {
-		columnType, isOk := columnsToTypes.GetColumn(column)
-		if isOk && columnType.ToastColumn {
-			if columnType.KindDetails == typing.Struct {
-				if bigQueryTypeCasting {
-					_columns = append(_columns,
-						fmt.Sprintf(`%s= CASE WHEN TO_JSON_STRING(cc.%s) != '{"key": "%s"}' THEN cc.%s ELSE c.%s END`,
-							// col CASE when TO_JSON_STRING(cc.col) != { 'key': TOAST_UNAVAILABLE_VALUE }
-							column, column, constants.ToastUnavailableValuePlaceholder,
-							// cc.col ELSE c.col END
-							column, column))
-				} else {
-					_columns = append(_columns,
-						fmt.Sprintf("%s= CASE WHEN cc.%s != {'key': '%s'} THEN cc.%s ELSE c.%s END",
-							// col CASE WHEN cc.col
-							column, column,
-							// { 'key': TOAST_UNAVAILABLE_VALUE } THEN cc.col ELSE c.col END",
-							constants.ToastUnavailableValuePlaceholder, column, column))
-				}
-			} else {
-				// t.column3 = CASE WHEN t.column3 != '__debezium_unavailable_value' THEN t.column3 ELSE s.column3 END
-				_columns = append(_columns,
-					fmt.Sprintf("%s= CASE WHEN cc.%s != '%s' THEN cc.%s ELSE c.%s END",
-						// col = CASE WHEN cc.col != TOAST_UNAVAILABLE_VALUE
-						column, column, constants.ToastUnavailableValuePlaceholder,
-						// THEN cc.col ELSE c.col END
-						column, column))
-			}
-
-		} else {
-			// This is to make it look like: objCol = cc.objCol
-			_columns = append(_columns, fmt.Sprintf("%s=cc.%s", column, column))
-		}
-
-	}
-
-	return strings.Join(_columns, ",")
 }
 
 // Empty will iterate over a list, if one of the item in the list is empty, it will return true
