+++ lib/array/strings_test.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/artie-labs/transfer/lib/typing"
-
 	"github.com/artie-labs/transfer/lib/config/constants"
 	"github.com/stretchr/testify/assert"
 )
@@ -74,97 +72,6 @@ func TestToArrayString(t *testing.T) {
 		actualString, actualErr := InterfaceToArrayString(testCase.val)
 		assert.Equal(t, testCase.expectedList, actualString, testCase.name)
 		assert.Equal(t, testCase.expectedErr, actualErr, testCase.name)
-	}
-
-}
-
-func TestColumnsUpdateQuery(t *testing.T) {
-	type testCase struct {
-		name           string
-		columns        []string
-		columnsToTypes typing.Columns
-		expectedString string
-		bigQuery       bool
-	}
-
-	fooBarCols := []string{"foo", "bar"}
-
-	var (
-		happyPathCols      typing.Columns
-		stringAndToastCols typing.Columns
-		lastCaseColTypes   typing.Columns
-	)
-	for _, col := range fooBarCols {
-		happyPathCols.AddColumn(typing.Column{
-			Name:        col,
-			KindDetails: typing.String,
-			ToastColumn: false,
-		})
-	}
-	for _, col := range fooBarCols {
-		var toastCol bool
-		if col == "foo" {
-			toastCol = true
-		}
-
-		stringAndToastCols.AddColumn(typing.Column{
-			Name:        col,
-			KindDetails: typing.String,
-			ToastColumn: toastCol,
-		})
-	}
-
-	lastCaseCols := []string{"a1", "b2", "c3"}
-
-	for _, lastCaseCol := range lastCaseCols {
-		kd := typing.String
-		var toast bool
-		// a1 - struct + toast, b2 - string + toast, c3 = regular string.
-		if lastCaseCol == "a1" {
-			kd = typing.Struct
-			toast = true
-		} else if lastCaseCol == "b2" {
-			toast = true
-		}
-
-		lastCaseColTypes.AddColumn(typing.Column{
-			Name:        lastCaseCol,
-			KindDetails: kd,
-			ToastColumn: toast,
-		})
-	}
-
-	testCases := []testCase{
-		{
-			name:           "happy path",
-			columns:        fooBarCols,
-			columnsToTypes: happyPathCols,
-			expectedString: "foo=cc.foo,bar=cc.bar",
-		},
-		{
-			name:           "string and toast",
-			columns:        fooBarCols,
-			columnsToTypes: stringAndToastCols,
-			expectedString: "foo= CASE WHEN cc.foo != '__debezium_unavailable_value' THEN cc.foo ELSE c.foo END,bar=cc.bar",
-		},
-		{
-			name:           "struct, string and toast string",
-			columns:        lastCaseCols,
-			columnsToTypes: lastCaseColTypes,
-			expectedString: "a1= CASE WHEN cc.a1 != {'key': '__debezium_unavailable_value'} THEN cc.a1 ELSE c.a1 END,b2= CASE WHEN cc.b2 != '__debezium_unavailable_value' THEN cc.b2 ELSE c.b2 END,c3=cc.c3",
-		},
-		{
-			name:           "struct, string and toast string (bigquery)",
-			columns:        lastCaseCols,
-			columnsToTypes: lastCaseColTypes,
-			bigQuery:       true,
-			expectedString: `a1= CASE WHEN TO_JSON_STRING(cc.a1) != '{"key": "__debezium_unavailable_value"}' THEN cc.a1 ELSE c.a1 END,b2= CASE WHEN cc.b2 != '__debezium_unavailable_value' THEN cc.b2 ELSE c.b2 END,c3=cc.c3`,
-		},
-	}
-
-	for _, _testCase := range testCases {
-		actualQuery := ColumnsUpdateQuery(_testCase.columns, _testCase.columnsToTypes, _testCase.bigQuery)
-		assert.Equal(t, _testCase.expectedString, actualQuery, _testCase.name)
 	}
 
 }
